Upgrade to new Homebridge Version
