+++ lib/solidmationApi.js
@@ -182,6 +182,12 @@ class Solidmation {
   async setDeviceStatus(DeviceID, mode) {
     console.log(`Set ${DeviceID}`, mode)
     const device = this.devices.find(device => device.DeviceID === DeviceID)
+
+    if (mode.mode == device.endpointValues.mode) {
+      console.log(`No required update ${DeviceID}`, mode)
+      return
+    }
+
     const payload = {
       ...device.endpointValues,
       ...mode,
