+++ package.json
@@ -3,7 +3,7 @@
   "keywords": [
     "homebridge-plugin"
   ],
-  "version": "1.0.0",
+  "version": "2.0.0",
   "description": "BGH Smart Control plugin for homebridge",
   "main": "dist/index.js",
   "author": "Sergio Behrends <sergio@behrends.com.ar>",
@@ -36,11 +36,12 @@
     "@babel/cli": "^7.2.3",
     "@babel/core": "^7.4.0",
     "@babel/preset-env": "^7.4.2",
+    "homebridge": "^1.1.0",
     "jest": "^24.5.0",
     "nock": "^10.0.6"
   },
   "engines": {
     "node": ">=0.12.0",
-    "homebridge": ">=0.2.0"
+    "homebridge": ">=1.1.0"
   }
 }
